+++ src/main/java/ir/ac/aut/hesabgar_group/manager/GroupManager.java
@@ -38,8 +38,8 @@ public class GroupManager {
         groupInfo.setCreationDate(new Date());
         groupInfo.setGroupName(creatingGroupInfoRequest.getGroupName());
 
-        Map<String, Integer> groupBalance = new HashMap<String, Integer>();
-        groupBalance.put(adminInfo.getId(), 0);
+        Map<String, Float> groupBalance = new HashMap<String, Float>();
+        groupBalance.put(adminInfo.getId(),(float) 0);
         groupInfo.setGroupBalance(groupBalance);
 
         List<GroupMember> groupMembers = new ArrayList<>();
@@ -93,8 +93,8 @@ public class GroupManager {
             memberInvAndDelEvents.add(new MemberInvAndDelEvent(newInvUser.getUserName(), true, new Date()));
             groupInfo.setMemberInvAndDelEvents(memberInvAndDelEvents);
 
-            Map<String, Integer> groupBalance = groupInfo.getGroupBalance();
-            groupBalance.put(newInvUser.getId(), 0);
+            Map<String, Float> groupBalance = groupInfo.getGroupBalance();
+            groupBalance.put(newInvUser.getId(), (float) 0);
             groupInfo.setGroupBalance(groupBalance);
 
             List<JoinedGroupInfo> newUserJoinedGroupInfos = newInvUser.getJoinedGroupList();
@@ -117,7 +117,7 @@ public class GroupManager {
         boolean deletable = false;
         if (adminUser.getId().equals(groupInfo.getAdmin())) {
             deletable = true;
-            Map<String, Integer> groupBalance = groupInfo.getGroupBalance();
+            Map<String, Float> groupBalance = groupInfo.getGroupBalance();
             for (String member : groupBalance.keySet()) {
                 groupUsers.add(member);
                 if (groupBalance.get(member) != 0) {
@@ -198,8 +198,8 @@ public class GroupManager {
         UserInfo adminUser = userInfoRepo.getUserInfoById(deletingGroupMemberRequest.getAdminId());
         UserInfo toBeDeletedUser = userInfoRepo.getUserInfoById(deletingGroupMemberRequest.getUserId());
         if (adminUser.getId().equals(groupInfo.getAdmin())) {
-            Map<String, Integer> groupBalance = groupInfo.getGroupBalance();
-            Map<String, Integer> newGroupBalance = new HashMap<>();
+            Map<String, Float> groupBalance = groupInfo.getGroupBalance();
+            Map<String, Float> newGroupBalance = new HashMap<>();
             if (groupBalance.get(toBeDeletedUser.getId()).equals(0)) {
                 for (String member : groupBalance.keySet()) {
                     if (!member.equals(toBeDeletedUser.getUserName())) {
@@ -241,8 +241,8 @@ public class GroupManager {
     public ResponseEntity<Object> addingInvoice(AddingInvoiceRequest addingInvoiceRequest) {
         //finding payment term and add it to groupMembers
         boolean allowed = false;
-        int invoiceAdminUserBalance = 0;
-        int sum = 0;
+        float invoiceAdminUserBalance = 0;
+        float sum = 0;
         GroupInfo groupInfo = groupInfoRepo.getGroupInfoById(addingInvoiceRequest.getGroupId());
         for (GroupMember groupMember : groupInfo.getMembers()) {
             if (groupMember.getUserId().equals(addingInvoiceRequest.getUserId()) && groupMember.isInvoiceAdmin()) {
@@ -260,8 +260,6 @@ public class GroupManager {
         if(sum * -1 != (addingInvoiceRequest.getTotalPaidValue() - invoiceAdminUserBalance)){
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
         }
-
-
         if (allowed) {
             //making a invoiceEvents and add it to group
             groupInfo = invoiceHelper.makingNewInvoiceEvent(groupInfo, addingInvoiceRequest);
